Removed nanoGUI, turned on all compiler warnings and solved all compiler warnings

+++ CMakeTemplateProject/include/PointOfInterest.cpp
@@ -2,7 +2,7 @@
 #include <vector>
 #include "PointOfInterest.h"
 
-PointOfInterest::PointOfInterest(int x, int y, pointOfInterestType type, bool obstacle) : x(x), y(y), type(type), isObstacle(obstacle) {
+PointOfInterest::PointOfInterest(size_t x, size_t y, pointOfInterestType type, bool obstacle) : x(x), y(y), type(type), isObstacle(obstacle) {
 	/**
 	Constructor for Point of Interest.
 	*/
